requirements.txt
torch==1.12.0
torchvision==0.13.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
Pillow>=8.0.0